Add checks for parsing integers in LEGATE_CONFIG (#3253)
Make all integer arguments to LEGATE_CONFIG signed integer types. Add appropriate checks to these args (either > 0 or >= 0).
In Config, use unsigned integer types for these arguments because we already checked them when parsing.

* WIP: add checks to flags

* fix build

* WIP: fix tests

* fix checks

* fix tests

* fix tidy errors. Add print for omps

* add print for new arg

* remove env var changes for a later PR

* resolve conflicts, fix build

+++ src/cpp/legate/partitioning/detail/restriction.cc
@@ -97,13 +97,13 @@ void Restrictions::join_inplace(const Restrictions& other)
 
 [[nodiscard]] std::tuple<SmallVector<std::size_t, LEGATE_MAX_DIM>,
                          SmallVector<std::uint32_t, LEGATE_MAX_DIM>,
-                         std::int64_t>
+                         std::uint64_t>
 Restrictions::prune_dimensions(Span<const std::uint64_t> shape) const
 {
   // Prune out any dimensions that are 1
   SmallVector<std::size_t, LEGATE_MAX_DIM> temp_shape{};
   SmallVector<std::uint32_t, LEGATE_MAX_DIM> temp_dims{};
-  std::int64_t volume = 1;
+  std::uint64_t volume = 1;
 
   temp_dims.reserve(shape.size());
   temp_shape.reserve(shape.size());
@@ -115,7 +115,7 @@ Restrictions::prune_dimensions(Span<const std::uint64_t> shape) const
     }
     temp_shape.push_back(extent);
     temp_dims.push_back(dim);
-    volume *= static_cast<std::int64_t>(extent);
+    volume *= extent;
   }
   return {std::move(temp_shape), std::move(temp_dims), volume};
 }
